Fix errors and update README

+++ src/main.rs
@@ -20,7 +20,7 @@ pub fn prove_sha256(msg: &[u8]) -> Result<()> {
     type C = PoseidonGoldilocksConfig;
     type F = <C as GenericConfig<D>>::F;
     let mut builder = CircuitBuilder::<F, D>::new(CircuitConfig::standard_recursion_config());
-    let targets = make_circuits(&mut builder, len as u128);
+    let targets = make_circuits(&mut builder, len as u64);
     let mut pw = PartialWitness::new();
 
     for i in 0..len {
